+++ src/pages/UpdatePasswordPage/UpdatePasswordPage.module.css
@@ -5,30 +5,11 @@
   min-height: 605px;
   }
   
-  .mainstr {
-    display: flex;
-    justify-content: center;
-  }
-
-  .form {
-    max-width: 280px;
-  }
-  
   @media screen and (min-width: 768px) {
     .container {
       padding: 40px 0 0 0;
       min-height: 704px;
     }
-
-    .form {
-      max-width: 680px;
-    }
-  
-    .mainstr {
-      justify-content: flex-start;
-    }
-
-   
   }
 
   @media screen and (min-width: 1440px) {
@@ -36,173 +17,4 @@
       padding: 49px 0 0 0;
       min-height: 740px;
     }
-  
-    .mainstr {
-      margin-top: 113px;
-      justify-content: flex-end;
-      margin-right: 90px;
-    }
-
   }
-
-  .formControl {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    flex-direction: column;
-    margin-bottom: 16px;
-    gap: 20px;
-  }
-  
-  .formTitle {
-    font-size: 26px;
-    margin-bottom: 16px;
-  }
-  
-  .formLabel {
-    margin-bottom: 8px;
-  }
-  
-  .stack {
-    position: relative;
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-  }
-  
-  .inputBox {
-    font-size: 16px;
-    color: var(--primary-color-blue);
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    padding: 12px 10px;
-    width: 280px;
-    border: 1px solid var(--secondary-color-blue-3);
-    border-radius: 6px;
-    outline: none;
-  }
-  
-  @media screen and (min-width: 768px) {
-    .inputBox {
-      width: 336px;
-    }
-  }
-  
-  @media screen and (min-width: 1440px) {
-    .inputBox {
-      width: 384px;
-    }
-  }
-
-  .inputBoxRep {
-    font-size: 16px;
-    color: var(--primary-color-blue);
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    padding: 12px 10px;
-    width: 280px;
-    border: 1px solid var(--secondary-color-blue-3);
-    border-radius: 6px;
-    outline: none;
-}
-
-@media screen and (min-width: 768px) {
-    .inputBoxRep {
-        width: 336px;
-    }
-}
-
-@media screen and (min-width: 1440px) {
-    .inputBoxRep {
-        width: 384px;
-    }
-}
-
-.iconeye {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  stroke: var(--primary-color-blue);
-}
-
-.inputpassword {
-  font-size: 16px;
-  color: var(--primary-color-blue);
-  outline: none;
-  border: none;
-  padding: 0;
-  width: 280px;
-}
-
-@media screen and (min-width: 768px) {
-    .inputpassword {
-        width: 336px;
-    }
-}
-
-@media screen and (min-width: 1440px) {
-    .inputpassword {
-        width: 384px;
-    }
-}
-
-.inputpassword::placeholder {
-    color: var(--secondary-color-blue-3);
-  }
-  
-  .button {
-    background-color: var(--primary-color-blue);
-    color: var(--primary-color-white);
-    border: none;
-    border-radius: 6px;
-    cursor: pointer;
-    width: 280px;
-    height: 44px;
-    font-size: 16px;
-    line-height: 20px;
-  }
-
-  @media screen and (min-width: 768px) {
-    .button {
-      width: 336px;
-      height: 44px;
-      font-size: 18px;
-      font-weight: 500;
-      line-height: 24px;
-    }
-  
-  }
-
-  @media screen and (min-width: 1440px) {
-    .button {
-      width: 384px;
-      height: 44px;
-  
-      font-size: 18px;
-      line-height: 24px;
-    }
-  }
-  
-  .errormessage {
-    position: absolute;
-    bottom: -18px;
-    left: 0;
-    color: var(--secondary-color-red);
-    font-size: 14px;
-  
-  }
-  
-  .errorBorder {
-    border: 1px solid var(--secondary-color-red);
-  }
-  
-  .errorInput {
-      color: var(--secondary-color-red);
-  }
-  
-  .errorInput::placeholder {
-      color: var(--secondary-color-red);
-  }
-
